Module Risk.Battle

val battle_simulator : Player.owned_territory -> Player.owned_territory -> Player.owned_territory

battle_simulator terr1 terr2 produces the winning country when a battle between terr1 and terr2 occurs, and it updates the amount of troops the winning country has

val odds_main : Player.owned_territory -> Player.owned_territory -> int -> float

odds_main terr1 terr2 is the chance of the attacking territory winning the battle in a battle between terr1 and terr2 This is mainly intended for testing purposes