Module Risk.Cards

type card =

| Troop

| Horse

| Cannon	(*
A card can be a troop, a horse, or a cannon
*)

val card_count : card list -> int list

card_count gives a count of each type of card that cards has, with the order being troops, horses, and cannons

val has_pattern : card list -> bool

has_pattern returns whether cards has a pattern

val pattern_troops : card list -> int

pattern_troops gives the troop bonus that cards produces

val must_trade_in : card list -> bool

must_trade_in says if a player must trade in their cards

val remove_cards : card list -> int -> card list

remove_cards takes cards from a card list depending on the amount of their bonus