Module Risk.Draft

val count_terrs : Player.owned_territory list -> int

count_terrs lst counts the number of territories in lst

val continent_bonus : Territories.continent list -> Player.player -> int

continent_bonus lst p gives the amount of bonus troops a player p get for holding continents based on lst

val allocate_troops : Player.player -> int

allocate_troops p gives the amount of troops that a player p gets to allocate to their territories at the start of their turn

val add_troops : Player.player -> Territories.territory -> int -> Player.player

add_troops p terr num adds given number num of troops to a territory terr owned by player p