Module Risk.Fortify

exception CantUseToFortify

val to_terr : Player.owned_territory list -> Territories.territory list

to_terr lst converts a player's owned_territory list lst to a territory list

val can_fortify : Player.player -> Territories.territory -> Territories.territory list -> Territories.territory list

can_fortify p terr lst is the list of territories that a player p can fortify fromm territory terr

val fortify : Player.player -> int -> Territories.territory -> Territories.territory -> Territories.territory list -> Player.player

fortify p troop terr1 terr2 moves a given number of troops (troop) from one territory terr1 to another terr2 in the fortify phase. troop must be greater than 1. Raises CantUseToFortify if terr1 cannot provide troops. Raises InvalidTerritory if a territory cannot be fortified