Module Risk.Init

type gamestate = {

p1 : Player.player;

p2 : Player.player;

p3 : Player.player;

p4 : Player.player;

state : int;

}

val fully_init_game : string -> string -> string -> string -> int -> Territories.territory list -> gamestate

val battle_init : gamestate -> Player.player -> Player.player -> Player.owned_territory -> Player.owned_territory -> int -> int -> gamestate

val list_available_borders : gamestate -> string list -> Player.player -> string

name_list game lst p lists the borders of a territory that the current player p can attack in game game

val find_owner_of_terr : gamestate -> string -> Player.player

find_owner_of_terr game terr finds the player who owns terr in game