Module Risk.Player

type owned_territory = {

place : Territories.territory;

troops : int;

}

An owned_territory contains a given territory and the troops in it

type player = {

name : string;

terrs : owned_territory list;

cards : Cards.card list;

}

A player has a name, a list of territories they own, and a list of their cards

val num_troops_in_terr : string -> player -> int

num_troops_in_terr terr p gives troops in terr owned by player p

val to_owned_terr : owned_territory list -> Territories.territory -> owned_territory

to_owned_terr lst terr returns the owned_territory of a given territory that a player owns terr

val terrs_to_string : owned_territory list -> string

terrs_to_string lst returns the owned territories list lst in a string with | in between

val name_list_owned : owned_territory list -> string list

name_list_owned lst converts an owned_territory list lst to a list of the names of the territories

val list_of_available_borders : string list -> player -> string list

name_list lst p makes a list of the borders of a territory that the current player can attack based on the borders lst and player p

val winner : player -> Territories.territory list -> bool

winner p lst says whether the given player p has won the game based on their list of territories lst

val lost : player -> bool

lost p says whether the given player p has lost

val show_cards : player -> string

show_cards p gives the string representation of a player p's cards

val give_card : player -> player

give_card p gives a player p a random card