Module Risk.Territories

type territory = {

name : string;

borders : string list;

}

A territory represents a country with a name and a list of bordering countries

type continent = {

parts : territory list;

bonus : int;

}

A continent is a list of territories and contains the troop bonus value that anyone who holds the country recieves at the start of their turn

val alaska : territory

The territories alaska, northwest_territory, ... , yakutsk are all of the territories on the game board

val northwest_territory : territory

val alberta : territory

val western_us : territory

val central_america : territory

val eastern_us : territory

val ontario : territory

val quebec : territory

val greenland : territory

val venezuela : territory

val peru : territory

val argentina : territory

val brazil : territory

val north_africa : territory

val congo : territory

val south_africa : territory

val egypt : territory

val madagascar : territory

val east_africa : territory

val western_europe : territory

val great_britain : territory

val iceland : territory

val scandinavia : territory

val northern_europe : territory

val southern_europe : territory

val ukraine : territory

val indonesia : territory

val new_guinea : territory

val western_australia : territory

val eastern_australia : territory

val middle_east : territory

val afghanistan : territory

val china : territory

val india : territory

val irkutsk : territory

val japan : territory

val kamchatka : territory

val mongolia : territory

val siam : territory

val siberia : territory

val ural : territory

val yakutsk : territory

val asia : continent

The continents in play on the game board as represented as asia, ... , north_america

val australia : continent

val europe : continent

val africa : continent

val south_america : continent

val north_america : continent

val all_territories : territory list

all_territories is a list of all of the territories

val all_continents : continent list

all_continents is a list of the continents

val is_border : territory -> territory -> bool

is_border says whether the two given territories border one another

val is_territory : string -> bool

is_territory says if name is the name of a valid territory

exception InvalidTerritory

val to_territory : string -> territory list -> territory

to_territory name list converts a name to a territory. Raises InvalidTerritory exception if there is no territory with name

val say_borders : territory -> string list

say_borders lists the borders of place

val sub_list : territory list -> territory list -> bool

sub_list says whether player owns a given continent_iq

val name_list_terrs : territory list -> string list

name_list_terrs converts an territory list to a list of the names of the territories

val continentsfake : territory list

continentsfake is a list of all of territories in test map